deduped some logic around checking round time
updated readme
